requirements.txt
pandas==1.2.3
numpy==1.20.1
pytest==6.2.2
scikit-learn==0.24.1
scipy==1.6.1
matplotlib==3.3.4
seaborn==0.11.1
dvc==2.0.5
joblib==1.0.1
xgboost==1.4.2
PyYAML==5.4.1
